+++ changelog.md
@@ -1,4 +1,5 @@
-- Custom AI: Number of improvements to collision avoidance logic. Ships are now less likely to bump into allies. 
+- Custom AI: Number of improvements to collision avoidance logic. Ships are now less likely to bump into allies.
+- Custom AI: Ships not avoid flying into mines.
 
 Version v2.2.1